requirements.txt
aiogram
apscheduler
openai
requests
python-dotenv
tinytag
pytz
